+++ backend/data/src/main/kotlin/io/tolgee/dtos/request/UserUpdatePasswordRequestDto.kt
@@ -5,7 +5,7 @@ import javax.validation.constraints.Size
 
 @JsonIgnoreProperties(ignoreUnknown = true)
 data class UserUpdatePasswordRequestDto(
-  @field:Size(min = 8, max = 50)
+  @field:Size(max = 50)
   var currentPassword: String = "",
 
   @field:Size(min = 8, max = 50)
