+++ webapp/src/component/security/Login/LoginCredentialsForm.tsx
@@ -1,21 +1,13 @@
-import React, {
-  FunctionComponent,
-  RefObject,
-  useEffect,
-  useRef,
-  useState,
-} from 'react';
+import React, { RefObject, useEffect } from 'react';
 import { Alert, Button, Link as MuiLink, Typography } from '@mui/material';
 import Box from '@mui/material/Box';
 import { T, useTranslate } from '@tolgee/react';
 import { useSelector } from 'react-redux';
-import { Link, Redirect, useHistory } from 'react-router-dom';
+import { Link } from 'react-router-dom';
 import { container } from 'tsyringe';
 
-import { components } from 'tg.service/apiSchema.generated';
 import { LINKS } from 'tg.constants/links';
 import { useConfig } from 'tg.globalContext/helpers';
-import { SecurityService } from 'tg.service/SecurityService';
 import { GlobalActions } from 'tg.store/global/GlobalActions';
 import { AppState } from 'tg.store/index';
 
@@ -31,24 +23,15 @@ import {
   OAuthService,
 } from 'tg.component/security/OAuthService';
 
-interface LoginProps {}
-
-type LoginRequestDto = components['schemas']['LoginRequest'];
-
 const globalActions = container.resolve(GlobalActions);
-const securityServiceIns = container.resolve(SecurityService);
 
 type Credentials = { username: string; password: string };
 type LoginViewCredentialsProps = {
   credentialsRef: RefObject<Credentials>;
   onMfaEnabled: () => void;
 };
-type LoginViewTotpProps = {
-  credentialsRef: RefObject<Credentials>;
-  onMfaCancel: () => void;
-};
 
-export function LoginViewCredentials(props: LoginViewCredentialsProps) {
+export function LoginCredentialsForm(props: LoginViewCredentialsProps) {
   const t = useTranslate();
   const remoteConfig = useConfig();
   const security = useSelector((state: AppState) => state.global.security);
@@ -188,107 +171,3 @@ export function LoginViewCredentials(props: LoginViewCredentialsProps) {
     </DashboardPage>
   );
 }
-
-export function LoginViewTotp(props: LoginViewTotpProps) {
-  const t = useTranslate();
-  const security = useSelector((state: AppState) => state.global.security);
-  const authLoading = useSelector(
-    (state: AppState) => state.global.authLoading
-  );
-
-  return (
-    <DashboardPage>
-      <CompactView
-        windowTitle={t('account-security-mfa')}
-        title={t('account-security-mfa')}
-        alerts={
-          security.loginErrorCode &&
-          !authLoading && (
-            <Alert severity="error">
-              <T>{security.loginErrorCode}</T>
-            </Alert>
-          )
-        }
-        content={
-          <StandardForm
-            initialValues={
-              {
-                username: props.credentialsRef.current!.username,
-                password: props.credentialsRef.current!.password,
-                otp: '',
-              } as LoginRequestDto
-            }
-            submitButtons={
-              <Box mt={2}>
-                <Box display="flex" flexDirection="column" alignItems="stretch">
-                  <LoadingButton
-                    loading={authLoading}
-                    variant="contained"
-                    color="primary"
-                    type="submit"
-                    data-cy="login-button"
-                  >
-                    <T>login_login_button</T>
-                  </LoadingButton>
-                </Box>
-              </Box>
-            }
-            onSubmit={(data) => globalActions.login.dispatch(data)}
-          >
-            <TextField
-              name="otp"
-              label={<T>account-security-mfa-otp-code</T>}
-              variant="standard"
-            />
-          </StandardForm>
-        }
-        footer={
-          <Box display="flex" justifyContent="flex-end">
-            <MuiLink onClick={() => props.onMfaCancel()} component={'button'}>
-              <Typography variant="caption">
-                <T>global_cancel_button</T>
-              </Typography>
-            </MuiLink>
-          </Box>
-        }
-      />
-    </DashboardPage>
-  );
-}
-
-// noinspection JSUnusedLocalSymbols
-export const LoginView: FunctionComponent<LoginProps> = (props) => {
-  const credentialsRef = useRef({ username: '', password: '' });
-  const [mfaRequired, setMfaRequired] = useState(false);
-
-  const security = useSelector((state: AppState) => state.global.security);
-  const remoteConfig = useConfig();
-  const history = useHistory();
-
-  if (history.location.state && (history.location.state as any).from) {
-    securityServiceIns.saveAfterLoginLink((history.location.state as any).from);
-  }
-
-  if (!remoteConfig.authentication || security.allowPrivate) {
-    return <Redirect to={LINKS.AFTER_LOGIN.build()} />;
-  }
-
-  if (mfaRequired) {
-    return (
-      <LoginViewTotp
-        credentialsRef={credentialsRef}
-        onMfaCancel={() => {
-          credentialsRef.current!.password = '';
-          setMfaRequired(false);
-        }}
-      />
-    );
-  }
-
-  return (
-    <LoginViewCredentials
-      credentialsRef={credentialsRef}
-      onMfaEnabled={() => setMfaRequired(true)}
-    />
-  );
-};
